requirements.txt
flask
psycopg2-binary
python-dotenv
ollama
customtkinter
requests